connectivity bind applies configuration overrides
