+++ components/kyma-environment-broker/internal/process/provisioning/connectivity_bind.go
@@ -1,11 +1,10 @@
 package provisioning
 
 import (
-	"encoding/json"
 	"fmt"
 	"time"
 
-	"github.com/pkg/errors"
+	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/process/connectivity_bind"
 
 	"github.com/google/uuid"
 	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal"
@@ -14,25 +13,6 @@ import (
 	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/storage"
 	"github.com/sirupsen/logrus"
 )
-
-type ConnectivityConfig struct {
-	ClientId            string `json:"clientid"`
-	ClientSecret        string `json:"clientsecret"`
-	ConnectivityService struct {
-		CAsPath        string `json:"CAs_path"`
-		CAsSigningPath string `json:"CAs_signing_path"`
-		ApiPath        string `json:"api_path"`
-		TunnelPath     string `json:"tunnel_path"`
-		Url            string `json:"url"`
-	} `json:"connectivity_service"`
-	SubaccountId                    string `json:"subaccount_id"`
-	SubaccountSubdomain             string `json:"subaccount_subdomain"`
-	TokenServiceDomain              string `json:"token_service_domain"`
-	TokenServiceUrl                 string `json:"token_service_url"`
-	TokenServiceUrlPattern          string `json:"token_service_url_pattern"`
-	TokenServiceUrlPatternTenantKey string `json:"token_service_url_pattern_tenant_key"`
-	Xsappname                       string `json:"xsappname"`
-}
 
 type ConnectivityBindStep struct {
 	operationManager *process.ProvisionOperationManager
@@ -74,7 +54,7 @@ func (s *ConnectivityBindStep) Run(operation internal.ProvisioningOperation, log
 		return s.operationManager.OperationFailed(operation, fmt.Sprintf("Connectivity provisioning failed: %s", resp.Description), log)
 	}
 	// execute binding
-	var connectivityOverrides *ConnectivityConfig
+	var connectivityOverrides *connectivity_bind.ConnectivityConfig
 	if !operation.Connectivity.Instance.Provisioned {
 		if operation.Connectivity.BindingID == "" {
 			operation.Connectivity.BindingID = uuid.New().String()
@@ -84,11 +64,11 @@ func (s *ConnectivityBindStep) Run(operation internal.ProvisioningOperation, log
 			return s.handleError(operation, err, log, fmt.Sprintf("Bind() call failed"))
 		}
 		// get overrides
-		connectivityOverrides, err = GetConnectivityCredentials(respBinding.Binding)
+		connectivityOverrides, err = connectivity_bind.GetConnectivityCredentials(respBinding.Binding)
 		if err != nil {
 			return s.handleError(operation, err, log, fmt.Sprintf("unable to load config"))
 		}
-		encryptedOverrides, err := EncryptConnectivityConfig(s.secretKey, connectivityOverrides)
+		encryptedOverrides, err := connectivity_bind.EncryptConnectivityConfig(s.secretKey, connectivityOverrides)
 		if err != nil {
 			return s.handleError(operation, err, log, fmt.Sprintf("unable to encrypt config"))
 		}
@@ -106,82 +86,24 @@ func (s *ConnectivityBindStep) Run(operation internal.ProvisioningOperation, log
 		operation = op
 	} else {
 		// get the credentials from encrypted string in operation.Connectivity.Instance.
-		connectivityOverrides, err = DecryptConnectivityConfig(s.secretKey, operation.Connectivity.Overrides)
+		connectivityOverrides, err = connectivity_bind.DecryptConnectivityConfig(s.secretKey, operation.Connectivity.Overrides)
 		if err != nil {
 			return s.handleError(operation, err, log, fmt.Sprintf("unable to decrypt configs"))
 		}
+		applyOverrides(connectivityOverrides, operation)
 	}
 
-	// TODO: Decide how we want to pass this data to the SKR. Currently,
-	//       credentials are prepared as a ConnectivityConfig structure.
-	//       See the github card - https://github.com/orgs/kyma-project/projects/6#card-56776111
-	//       ...
-	//       - [ ] define what changes need to be done in KEB to
-	//             allow passing secrets data to the Provisioner
 	log.Debugf("Got Connectivity Service credentials from the binding.")
 
 	return operation, 0, nil
 }
 
+func applyOverrides(connectivityOverrides *connectivity_bind.ConnectivityConfig, operation internal.ProvisioningOperation) {
+	overrides := connectivity_bind.PrepareOverrides(connectivityOverrides)
+	operation.InputCreator.AppendOverrides(connectivity_bind.ConnectivityProxyComponentName, overrides)
+}
+
 func (s *ConnectivityBindStep) handleError(operation internal.ProvisioningOperation, err error, log logrus.FieldLogger, msg string) (internal.ProvisioningOperation, time.Duration, error) {
 	log.Errorf("%s: %s", msg, err)
 	return s.operationManager.OperationFailed(operation, msg, log)
-}
-
-func GetConnectivityCredentials(binding servicemanager.Binding) (*ConnectivityConfig, error) {
-	credentials := binding.Credentials
-	csMap, ok := credentials["connectivity_service"].(map[string]interface{})
-	if !ok {
-		return nil, fmt.Errorf(
-			"failed to convert connectivity_service part of the credentials to map[string]interface{}")
-	}
-
-	return &ConnectivityConfig{
-		ClientId:     credentials["clientid"].(string),
-		ClientSecret: credentials["clientsecret"].(string),
-		ConnectivityService: struct {
-			CAsPath        string `json:"CAs_path"`
-			CAsSigningPath string `json:"CAs_signing_path"`
-			ApiPath        string `json:"api_path"`
-			TunnelPath     string `json:"tunnel_path"`
-			Url            string `json:"url"`
-		}{
-			CAsPath:        csMap["CAs_path"].(string),
-			CAsSigningPath: csMap["CAs_signing_path"].(string),
-			ApiPath:        csMap["CAs_signing_path"].(string),
-			TunnelPath:     csMap["tunnel_path"].(string),
-			Url:            csMap["url"].(string),
-		},
-		SubaccountId:                    credentials["subaccount_id"].(string),
-		SubaccountSubdomain:             credentials["subaccount_subdomain"].(string),
-		TokenServiceDomain:              credentials["token_service_domain"].(string),
-		TokenServiceUrl:                 credentials["token_service_url"].(string),
-		TokenServiceUrlPattern:          credentials["token_service_url_pattern"].(string),
-		TokenServiceUrlPatternTenantKey: credentials["token_service_url_pattern_tenant_key"].(string),
-		Xsappname:                       credentials["xsappname"].(string),
-	}, nil
-}
-
-func EncryptConnectivityConfig(secretKey string, overrides *ConnectivityConfig) (string, error) {
-	marshalledOverrides, err := json.Marshal(*overrides)
-	if err != nil {
-		return "", errors.Wrap(err, "while encoding connectivity overrides")
-	}
-	encryptedOverrides, err := storage.NewEncrypter(secretKey).Encrypt(marshalledOverrides)
-	if err != nil {
-		return "", errors.Wrap(err, "while encrypting connectivity overrides")
-	}
-	return string(encryptedOverrides), nil
-}
-
-func DecryptConnectivityConfig(secretKey string, encryptedOverrides string) (*ConnectivityConfig, error) {
-	decryptedOverrides, err := storage.NewEncrypter(secretKey).Decrypt([]byte(encryptedOverrides))
-	if err != nil {
-		return nil, errors.Wrap(err, "while decrypting connectivity overrides")
-	}
-	connectivityOverrides := ConnectivityConfig{}
-	if err := json.Unmarshal(decryptedOverrides, &connectivityOverrides); err != nil {
-		return nil, errors.Wrap(err, "while unmarshalling connectivity overrides")
-	}
-	return &connectivityOverrides, nil
 }
